Fixed Buttons (For now)
I think I fixed the buttons for now. But if I wanted to add more buttons in the future, it wouldn't be as simple as just adding the buttons. I would have to change the media queries and a few other lines to make it work properly. I will try to fix this in the future but for now I think it works.

The current state of the body-div-1 is just positioning it relative to the other objects and aligning the top and bottom elements to the top and bottom. Then I made the height 100% and made up the space with margins.

It works with zooming (up until a certain point). To solve this issue I am thinking of making the footer a popup. However, I need to do some research to see if this works well with SEO (Search Engine Optimization).