Added a few things (Mobilemenu)
Added animation to button.
Menu actually works, and javascript backend is a lot more simple so it handles the opening and closing of it properly and efficiently (swapping of the display style from "none" to "flex" and vis versa)

Current bug, I need to fix the menu to not scroll but to size down when the screen size is smaller.

What I want to add is a slide down of the dropdown menu to notify to people that the menu is open. More intuitive clues is key to making that work properly.